Created a separate context for the WebSockets
